+++ examples/resources/iosxe_bgp_address_family_ipv4/resource.tf
@@ -3,7 +3,7 @@ resource "iosxe_bgp_address_family_ipv4" "example" {
   af_name                             = "unicast"
   ipv4_unicast_redistribute_connected = true
   ipv4_unicast_redistribute_static    = true
-  ipv4_unicast_aggregate_address = [
+  ipv4_unicast_aggregate_addresses = [
     {
       ipv4_address = "10.0.0.0"
       ipv4_mask    = "255.255.0.0"