+++ internal/provider/data_source_iosxe_bgp_address_family_ipv4_test.go
@@ -29,8 +29,8 @@ func TestAccDataSourceIosxeBGPAddressFamilyIPv4(t *testing.T) {
 	var checks []resource.TestCheckFunc
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_redistribute_connected", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_redistribute_static", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_aggregate_address.0.ipv4_address", "10.0.0.0"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_aggregate_address.0.ipv4_mask", "255.255.0.0"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_aggregate_addresses.0.ipv4_address", "10.0.0.0"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_aggregate_addresses.0.ipv4_mask", "255.255.0.0"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_networks_mask.0.network", "12.0.0.0"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_networks_mask.0.mask", "255.255.0.0"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_networks_mask.0.route_map", "RM1"))
@@ -67,7 +67,7 @@ func testAccDataSourceIosxeBGPAddressFamilyIPv4Config() string {
 	config += `	af_name = "unicast"` + "\n"
 	config += `	ipv4_unicast_redistribute_connected = true` + "\n"
 	config += `	ipv4_unicast_redistribute_static = true` + "\n"
-	config += `	ipv4_unicast_aggregate_address = [{` + "\n"
+	config += `	ipv4_unicast_aggregate_addresses = [{` + "\n"
 	config += `		ipv4_address = "10.0.0.0"` + "\n"
 	config += `		ipv4_mask = "255.255.0.0"` + "\n"
 	config += `	}]` + "\n"