+++ internal/provider/data_source_iosxe_bgp_address_family_ipv4_vrf.go
@@ -88,7 +88,7 @@ func (d *BGPAddressFamilyIPv4VRFDataSource) Schema(ctx context.Context, req data
 							MarkdownDescription: "Connected",
 							Computed:            true,
 						},
-						"ipv4_unicast_aggregate_address": schema.ListNestedAttribute{
+						"ipv4_unicast_aggregate_addresses": schema.ListNestedAttribute{
 							MarkdownDescription: "Configure BGP aggregate entries",
 							Computed:            true,
 							NestedObject: schema.NestedAttributeObject{