+++ internal/provider/resource_iosxe_bgp_address_family_ipv4_test.go
@@ -30,8 +30,8 @@ func TestAccIosxeBGPAddressFamilyIPv4(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "af_name", "unicast"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_redistribute_connected", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_redistribute_static", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_aggregate_address.0.ipv4_address", "10.0.0.0"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_aggregate_address.0.ipv4_mask", "255.255.0.0"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_aggregate_addresses.0.ipv4_address", "10.0.0.0"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_aggregate_addresses.0.ipv4_mask", "255.255.0.0"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_networks_mask.0.network", "12.0.0.0"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_networks_mask.0.mask", "255.255.0.0"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_address_family_ipv4.test", "ipv4_unicast_networks_mask.0.route_map", "RM1"))
@@ -84,7 +84,7 @@ func testAccIosxeBGPAddressFamilyIPv4Config_all() string {
 	config += `	af_name = "unicast"` + "\n"
 	config += `	ipv4_unicast_redistribute_connected = true` + "\n"
 	config += `	ipv4_unicast_redistribute_static = true` + "\n"
-	config += `	ipv4_unicast_aggregate_address = [{` + "\n"
+	config += `	ipv4_unicast_aggregate_addresses = [{` + "\n"
 	config += `		ipv4_address = "10.0.0.0"` + "\n"
 	config += `		ipv4_mask = "255.255.0.0"` + "\n"
 	config += `	}]` + "\n"